% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas_from_model.R
\name{betas_from_glm}
\alias{betas_from_glm}
\title{Extract model information from glm models}
\usage{
betas_from_glm(glm_model, intercept = FALSE)
}
\arguments{
\item{glm_model}{A fitted glm model object}

\item{intercept}{Whether to include the intercept in the output}
}
\value{
A data.table with the following columns:
\itemize{
 \item \code{predictor}: The name of the predictor
\item \code{beta}: The beta coefficient
\item \code{se_beta}: The standard error of the beta coefficient
\item \code{p_value}: The p-value of the beta coefficient
\item \code{or}: The odds ratio
\item \code{or_lower}: The lower bound of the odds ratio
\item \code{or_upper}: The upper bound of the odds ratio
}
}
\description{
Extract model information from glm models
}
