% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_weights.R
\name{cv_folds}
\alias{cv_folds}
\title{Cross-validation folds}
\usage{
cv_folds(data, k, weights, seed = NULL, strata = NULL, cluster = NULL, R = 1)
}
\arguments{
\item{data}{A data frame with information of cluster and strata indicators and sampling weights.}

\item{k}{Number of folds}

\item{weights}{Name of the column indicating sampling weights}

\item{seed}{seed}

\item{strata}{Name of the column indicating strata}

\item{cluster}{Name of the column indicating clusters}

\item{R}{Number of times the sample is split}
}
\value{
This function returns a new data frame with new columns, each of them indicating the fold of each unit.
}
\description{
This function allows calculating cross-validation folds.
}
