% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_f.R
\name{error_f}
\alias{error_f}
\title{Error function}
\usage{
error_f(
  data,
  l_yhat,
  method,
  cv_error_ind = c(TRUE, FALSE),
  R = NULL,
  k = NULL,
  B = NULL,
  outcome,
  family,
  weights = NULL
)
}
\arguments{
\item{data}{Data frame with the data}

\item{l_yhat}{List of matrices with the predicted values}

\item{method}{Method used to generate the replicate weights}

\item{cv_error_ind}{Logical indicating if the cross-validation error should be calculated}

\item{R}{Number of replicate weights}

\item{k}{Number of folds}

\item{B}{Number of bootstrap samples}

\item{outcome}{Name of the outcome variable}

\item{family}{Family of the outcome variable}

\item{weights}{Name of the weights variable}
}
\value{
Matrix with the error for each lambda
}
\description{
Error function
}
