% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_weights.R
\name{f_folds}
\alias{f_folds}
\title{Folds for cross-validation}
\usage{
f_folds(data, k = 5, strata = NULL, cluster = NULL, seed = NULL)
}
\arguments{
\item{data}{A data frame with information of cluster and strata indicators and sampling weights.}

\item{k}{Number of folds}

\item{strata}{Name of the column indicating strata}

\item{cluster}{Name of the column indicating clusters}

\item{seed}{seed}
}
\value{
This function returns a vector with the fold of each unit.
}
\description{
This function allows calculating folds for cross-validation.
}
