% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bin_diagnostics.R
\name{get_bin_diagnostics}
\alias{get_bin_diagnostics}
\title{Get suite of diagnostics for a risk score and a binary outcome}
\usage{
get_bin_diagnostics(
  data,
  outcome,
  exposure,
  covs = NULL,
  pctile_or = TRUE,
  pl = FALSE
)
}
\arguments{
\item{data}{A data frame}

\item{outcome}{The name of the outcome column}

\item{exposure}{The name of the exposure column}

\item{covs}{A vector of covariates}

\item{pctile_or}{Whether to calculate percentile-based ORs (default: TRUE)}

\item{pl}{Whether to use profile likelihood (default: FALSE)}
}
\value{
A data table with the following columns:
\itemize{
 \item \code{stat}: The name of the statistic
 \item \code{value}: The value of the statistic
}
}
\description{
Get suite of diagnostics for a risk score and a binary outcome
}
