% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bin_diagnostics.R
\name{getTopEffects}
\alias{getTopEffects}
\title{Helper: get top effects}
\usage{
getTopEffects(prob, data, outcome = "case", exposure, covs = NULL, pl = FALSE)
}
\arguments{
\item{prob}{A probability representing a percentile}

\item{data}{A data frame}

\item{outcome}{The name of the outcome column}

\item{exposure}{The name of the exposure column}

\item{covs}{A vector of covariates}

\item{pl}{Whether to use profile likelihood (default: FALSE)}
}
\description{
Helper: get top effects
}
