% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bin_diagnostics.R
\name{getTopPower}
\alias{getTopPower}
\title{Helper: get top power}
\usage{
getTopPower(prob, data, outcome, exposure)
}
\arguments{
\item{prob}{A probability representing a percentile}

\item{data}{A data frame}

\item{outcome}{The name of the outcome column}

\item{exposure}{The name of the exposure column}
}
\description{
Helper: get top power
}
