% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bin_diagnostics.R
\name{getValues}
\alias{getValues}
\title{Helper: get value}
\usage{
getValues(fitAsso, exposure)
}
\arguments{
\item{fitAsso}{A logistf object; used within `getTopEffects`}

\item{exposure}{The name of the exposure column}
}
\description{
Helper: get value
}
