% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_f.R
\name{loss_f}
\alias{loss_f}
\title{Loss function}
\usage{
loss_f(y_est, y, family = "binomial")
}
\arguments{
\item{y_est}{Estimated values}

\item{y}{Observed values}

\item{family}{Family of the outcome variable}
}
\value{
Vector with the loss for each observation
}
\description{
Loss function
}
