% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nagelkerke_r2.R
\name{nagelkerke_r2}
\alias{nagelkerke_r2}
\title{Helper: modified rcompanion::nagelkerke}
\usage{
nagelkerke_r2(fit, null = NULL, restrictNobs = FALSE)
}
\arguments{
\item{fit}{A fitted model object}

\item{null}{A fitted null model object}

\item{restrictNobs}{Whether to restrict the number of observations to the number of observations in the fitted model (default: FALSE)}
}
\description{
Helper: modified rcompanion::nagelkerke
}
