% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_fit.R
\name{plot_glmnet_vip}
\alias{plot_glmnet_vip}
\title{Plot variable importance (betas) from glmnet model}
\usage{
plot_glmnet_vip(x, exposure_var = label, beta_var = rel_beta, top_n = 15)
}
\arguments{
\item{x}{A glmnet model object}

\item{exposure_var}{The name of the exposure variable}

\item{beta_var}{The name of the beta variable to plot}

\item{top_n}{The number of top variables to plot}
}
\value{
A ggplot2 object
}
\description{
Plot variable importance (betas) from glmnet model
}
