% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_fit.R
\name{plot_ranger_vip}
\alias{plot_ranger_vip}
\title{Plot variable importance from ranger model}
\usage{
plot_ranger_vip(
  x,
  exposure_var = label,
  top_n = 15,
  caption = ggplot2::waiver()
)
}
\arguments{
\item{x}{Table of variable importance from ranger model (see tidy_ranger_imp)}

\item{exposure_var}{The name of the exposure variable}

\item{top_n}{The number of top variables to plot}

\item{caption}{Plot caption}
}
\value{
A ggplot2 object
}
\description{
Plot variable importance from ranger model
}
