% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_weights.R
\name{repl_weights_test}
\alias{repl_weights_test}
\title{Replicate weights for testing set}
\usage{
repl_weights_test(
  data,
  folds,
  test_fold,
  weights,
  strata = NULL,
  cluster = NULL
)
}
\arguments{
\item{data}{A data frame with information of cluster and strata indicators and sampling weights.}

\item{folds}{Name of the column indicating the fold of each unit.}

\item{test_fold}{Fold of the testing set.}

\item{weights}{Name of the column indicating sampling weights}

\item{strata}{Name of the column indicating strata}

\item{cluster}{Name of the column indicating clusters}
}
\value{
This function returns a vector with the replicate weights for the testing set.
}
\description{
This function allows calculating replicate weights for testing set.
}
