% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_weights.R
\name{replicate_sample}
\alias{replicate_sample}
\title{Replicate sample}
\usage{
replicate_sample(
  data,
  set,
  tag,
  strata,
  weights,
  r = 1,
  boot_type = c("bootstrap", "subbootstrap")
)
}
\arguments{
\item{data}{A data frame with information of cluster and strata indicators and sampling weights.}

\item{set}{Name of the column indicating the fold of each unit.}

\item{tag}{Fold of the testing set.}

\item{strata}{Name of the column indicating strata}

\item{weights}{Name of the column indicating sampling weights}

\item{r}{Number of times the sample is split}

\item{boot_type}{To choose between \code{bootstrap} or \code{subbootstrap}}
}
\value{
This function returns a new data frame with new columns, each of them indicating replicate weights for different subsets.
}
\description{
This function allows calculating replicate sample.
}
