% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_weights.R
\name{replicate_weights}
\alias{replicate_weights}
\title{Replicate weights (based on wlasso::replicate.weights())}
\usage{
replicate_weights(
  data,
  method = "bootstrap",
  cluster = NULL,
  strata = NULL,
  weights = NULL,
  k = 10,
  B = 200,
  R = 1,
  train_prob = NULL,
  method_split = c("dCV", "bootstrap", "subbootstrap"),
  seed = NULL
)
}
\arguments{
\item{data}{A data frame with information of cluster and strata indicators and sampling weights.}

\item{method}{To choose between \code{JKn}, \code{dCV}, \code{bootstrap}, \code{subbootstrap}, \code{BRR}, \code{split} or \code{extrapolation}}

\item{cluster}{Name of the column indicating clusters}

\item{strata}{Name of the column indicating strata}

\item{weights}{Name of the column indicating sampling weights}

\item{k}{Number of folds, if \code{dCV} method is selected}

\item{B}{Number of bootstrap resamples if \code{bootstrap} or \code{subbootstrap} methods are selected}

\item{R}{Number of times the sample is split (only for \code{cv}, \code{split} and \code{extrapolation} methods)}

\item{train_prob}{Proportion of PSUs set as training set.}

\item{method_split}{To choose between \code{dCV}, \code{bootstrap} or \code{subbootstrap}}

\item{seed}{seed}
}
\value{
This function returns a new data frame with new columns, each of them indicating replicate weights for different subsets.
}
\description{
This function allows calculating replicate weights.
}
