% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_weights.R
\name{rw_split}
\alias{rw_split}
\title{Split-sample}
\usage{
rw_split(
  data,
  train_prob,
  method = c("dCV", "bootstrap", "subbootstrap"),
  weights,
  strata = NULL,
  cluster = NULL,
  R = 1,
  seed = 1
)
}
\arguments{
\item{data}{A data frame with information of cluster and strata indicators and sampling weights.}

\item{train_prob}{Proportion of PSUs set as training set.}

\item{method}{To choose between \code{dCV}, \code{bootstrap} or \code{subbootstrap}}

\item{weights}{Name of the column indicating sampling weights}

\item{strata}{Name of the column indicating strata}

\item{cluster}{Name of the column indicating clusters}

\item{R}{Number of times the sample is split}

\item{seed}{seed}
}
\value{
This function returns a new data frame with new columns, each of them indicating replicate weights for different subsets.
}
\description{
This function allows calculating split-sample.
}
