% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_weights.R
\name{split_sample}
\alias{split_sample}
\title{Split-sample}
\usage{
split_sample(data, train_prob, r, strata = NULL, cluster = NULL, seed = 1)
}
\arguments{
\item{data}{A data frame with information of cluster and strata indicators and sampling weights.}

\item{train_prob}{Proportion of PSUs set as training set.}

\item{r}{Number of times the sample is split}

\item{strata}{Name of the column indicating strata}

\item{cluster}{Name of the column indicating clusters}

\item{seed}{seed}
}
\value{
This function returns a new data frame with new columns, each of them indicating the fold of each unit.
}
\description{
This function allows calculating split-sample.
}
