% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_weights.R
\name{split_strata}
\alias{split_strata}
\title{Extrapolation}
\usage{
split_strata(data, train_prob, strata = NULL, weights, seed = 1, R = 1)
}
\arguments{
\item{data}{A data frame with information of cluster and strata indicators and sampling weights.}

\item{train_prob}{Proportion of PSUs set as training set.}

\item{strata}{Name of the column indicating strata}

\item{weights}{Name of the column indicating sampling weights}

\item{seed}{seed}

\item{R}{Number of times the sample is split}
}
\value{
This function returns a new data frame with new columns, each of them indicating replicate weights for different subsets.
}
\description{
This function allows calculating extrapolation.
}
