% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_fit.R
\name{tidy_glmnet}
\alias{tidy_glmnet}
\title{Fit a glmnet model and extract variable importance (betas) and plot}
\usage{
tidy_glmnet(data, exposures, outcome, lambda, alpha, family = "binomial", ...)
}
\arguments{
\item{data}{A data frame}

\item{exposures}{A character vector of exposure variables}

\item{outcome}{A character vector of outcome variables}

\item{lambda}{A numeric value for the lambda parameter}

\item{alpha}{A numeric value for the alpha parameter}

\item{family}{A character value for the glmnet family argument}

\item{...}{Additional arguments to pass to glmnet}
}
\value{
List of objects: glmnet object, data table of betas, ggplot2 object
}
\description{
Fit a glmnet model and extract variable importance (betas) and plot
}
