% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_fit.R
\name{tidy_glmnet_betas}
\alias{tidy_glmnet_betas}
\title{Extract variable importance (betas) from glmnet model}
\usage{
tidy_glmnet_betas(x, phecode_description = TRUE)
}
\arguments{
\item{x}{A glmnet model object}

\item{phecode_description}{A logical indicating whether to include the phecode description}
}
\value{
A data table with the exposure and beta
}
\description{
Extract variable importance (betas) from glmnet model
}
