% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_fit.R
\name{tidy_ranger}
\alias{tidy_ranger}
\title{Fit a ranger model and extract variable importance and plot}
\usage{
tidy_ranger(
  formula,
  data,
  num_trees = 500,
  mtry,
  min.node.size,
  num_threads = 12,
  importance = "permutation",
  write_forest = TRUE,
  exposure_var = label,
  ...
)
}
\arguments{
\item{formula}{A formula}

\item{data}{A data frame}

\item{num_trees}{A numeric value for the number of trees}

\item{mtry}{A numeric value for the mtry parameter}

\item{min.node.size}{A numeric value for the min.node.size parameter}

\item{num_threads}{A numeric value for the num_threads parameter}

\item{importance}{A character value for the importance parameter}

\item{write_forest}{A logical value for the write_forest parameter}

\item{exposure_var}{The name of the exposure variable}

\item{...}{Additional arguments to pass to ranger}
}
\value{
List of objects: ranger object, data table of importance, ggplot2 object
}
\description{
Fit a ranger model and extract variable importance and plot
}
