% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_fit.R
\name{tidy_ranger_imp}
\alias{tidy_ranger_imp}
\title{Extract variable importance from ranger model}
\usage{
tidy_ranger_imp(x, phecode_description = TRUE)
}
\arguments{
\item{x}{A ranger model object}

\item{phecode_description}{A logical indicating whether to include the phecode description}
}
\value{
A data table with the exposure and importance
}
\description{
Extract variable importance from ranger model
}
