% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_fit.R
\name{tidy_wlasso}
\alias{tidy_wlasso}
\title{Fit a weighted lasso model and extract variable importance (betas) and plot}
\usage{
tidy_wlasso(
  data,
  exposures,
  outcome,
  weight_var = NULL,
  lambda = NULL,
  family = "binomial",
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{exposures}{A character vector of exposure variables}

\item{outcome}{A character vector of outcome variables}

\item{weight_var}{Name of the weight variable}

\item{lambda}{A numeric value for the lambda parameter}

\item{family}{A character value for the glmnet family argument}

\item{...}{Additional arguments to pass to glmnet}
}
\value{
List of objects: glmnet object, data table of betas, ggplot2 object
}
\description{
Fit a weighted lasso model and extract variable importance (betas) and plot
}
