% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_glmnet}
\alias{tune_glmnet}
\title{Tune lambda (and alpha) hyperparameters for lasso, ridge, and elastic net regression via glmnet}
\usage{
tune_glmnet(
  data,
  outcome,
  exposures,
  alpha,
  parallel = TRUE,
  n_folds = 10,
  family = "binomial",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{dataset}

\item{outcome}{name of outcome variable}

\item{exposures}{vector of exposure variable names; if NULL, assumes all non-outcome variables}

\item{alpha}{0 for lasso, 1 for ridge, or a vector of numeric values to try for elastic net}

\item{parallel}{parallelize?}

\item{n_folds}{number of folds for cross-validation}

\item{family}{model family (binomial or guassian)}

\item{verbose}{print additional information}

\item{...}{additional arguments passed to cv.glmnet}
}
\value{
return a table with hyperparameters and their values
}
\description{
Tune lambda (and alpha) hyperparameters for lasso, ridge, and elastic net regression via glmnet
}
