% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_models}
\alias{tune_models}
\title{Tune hyperparameters for (un)weighted regularized regression and unweighted random forest}
\usage{
tune_models(
  data,
  outcome,
  exposures = NULL,
  weight = NULL,
  parallel = TRUE,
  methods = c("ridge", "lasso", "enet", "rf"),
  n_folds = 10,
  n_cores = 1,
  alpha = NULL,
  mtry_seq = NULL,
  node_size_seq = NULL,
  n_trees = 500
)
}
\arguments{
\item{data}{dataset}

\item{outcome}{name of outcome variable}

\item{exposures}{vector of exposure variable names; if NULL, assumes all non-outcome variables}

\item{weight}{name of weight variable}

\item{parallel}{parallelize?}

\item{methods}{vector of method names to consider: ridge, lasso, enet, rf}

\item{n_folds}{number of folds for cross-validation}

\item{n_cores}{number of cores to use for parallelization}

\item{alpha}{value(s) of alpha to pass to tune_glmnet if ridge, lasso, and/or enet are selected}

\item{mtry_seq}{values of mtry to try via tune_rf if rf method is selected}

\item{node_size_seq}{value of min.node.size to try via tune_rf if rf method is selected}

\item{n_trees}{number of trees to grow if rf method is selected}
}
\value{
return a table with hyperparameters and their values
}
\description{
Tune hyperparameters for (un)weighted regularized regression and unweighted random forest
}
