% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_ranger}
\alias{tune_ranger}
\title{Tune mtry and min.node.size parameters for random forest via ranger}
\usage{
tune_ranger(
  data,
  outcome,
  exposures = NULL,
  mtry_seq = NULL,
  node_size_seq = NULL,
  n_cores = 1,
  n_trees = 500
)
}
\arguments{
\item{data}{dataset}

\item{outcome}{name of outcome variable}

\item{exposures}{vector of exposure variable names; if NULL, assumes all non-outcome variables}

\item{mtry_seq}{values of mtry to search over. defaults to some recommendations from Breiman}

\item{node_size_seq}{values of min.node.size to search over. defaults to 5 values between 3 and nrow(data)/100}

\item{n_cores}{number of cores if parallel}

\item{n_trees}{number of trees to grow}
}
\value{
return a table with hyperparameters and their values
}
\description{
Tune mtry and min.node.size parameters for random forest via ranger
}
