% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_wglmnet}
\alias{tune_wglmnet}
\title{Tune lambda parameter for weighted lasso via wlasso::wlasso (then fall back to glmnet if issues)}
\usage{
tune_wglmnet(
  data,
  outcome,
  exposures,
  alpha = 1,
  weight,
  n_folds = 10,
  parallel = parallel,
  verbose = TRUE,
  method = "dCV",
  family = "binomial",
  ...
)
}
\arguments{
\item{data}{dataset}

\item{outcome}{name of outcome variable}

\item{exposures}{vector of exposure variable names; if NULL, assumes all non-outcome variables}

\item{alpha}{0 for lasso, 1 for ridge, or a vector of numeric values to try for elastic net}

\item{weight}{name of weight variable}

\item{n_folds}{number of folds for cross-validation}

\item{parallel}{parallelize?}

\item{verbose}{print additional information}

\item{method}{method for replicate weight sampling}

\item{family}{model family (binomial or guassian)}

\item{...}{additional arguments passed to wglmnet}
}
\value{
return a table with hyperparameters and their values
}
\description{
Tune lambda parameter for weighted lasso via wlasso::wlasso (then fall back to glmnet if issues)
}
