% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune_wlasso}
\alias{tune_wlasso}
\title{Tune lambda parameter for weighted lasso via wlasso::wlasso (then fall back to glmnet if issues)}
\usage{
tune_wlasso(data, outcome, exposures, weight, n_folds = 10)
}
\arguments{
\item{data}{dataset}

\item{outcome}{name of outcome variable}

\item{exposures}{vector of exposure variable names; if NULL, assumes all non-outcome variables}

\item{weight}{name of weight variable}

\item{n_folds}{number of folds for cross-validation}
}
\value{
return a table with hyperparameters and their values
}
\description{
Tune lambda parameter for weighted lasso via wlasso::wlasso (then fall back to glmnet if issues)
}
