% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglmnet.R
\name{wglmnet}
\alias{wglmnet}
\title{Weighted glmnet prediction models using replicate weights (modified version of wlasso::wlasso())}
\usage{
wglmnet(
  data = NULL,
  outcome = NULL,
  predictors = NULL,
  cluster = NULL,
  strata = NULL,
  weights = NULL,
  design = NULL,
  family = "binomial",
  lambda_grid = NULL,
  alpha = 1,
  method = "bootstrap",
  k = 10,
  R = 1,
  B = 200,
  cv_error_ind = FALSE,
  train_prob = 0.7,
  method_split = c("dCV", "bootstrap", "subbootstrap"),
  seed = NULL
)
}
\arguments{
\item{data}{A data frame with information about the outcome variable and predictors and sampling weights.}

\item{outcome}{column of the outcome variable. It could be indicated as the column number or the column name.}

\item{predictors}{vector of columns indicating the covariates. It could be defined by numbers or names of the columns.}

\item{cluster}{name of the column indicating clusters}

\item{strata}{name of the column indicating strata}

\item{weights}{name of the column indicating sampling weights}

\item{design}{an object of class \code{survey.design}. It could be \code{NULL} if information about \code{cluster}, \code{strata}, \code{weights} and \code{data} are given.}

\item{family}{family to fit glmnet models, choose between \code{gaussian} or \code{binomial} (default).}

\item{lambda_grid}{a grid for penalization parameters. If it is not defined by the user, the function will define one.}

\item{alpha}{the elastic net mixing parameter, with \code{alpha=0} equivalent to ridge regression and \code{alpha=1} equivalent to lasso. Default is \code{alpha=1}.}

\item{method}{method to be applied to define replicate weights, to choose between one of these: \code{JKn}, \code{dCV}, \code{bootstrap} (default), \code{subbootstrap}, \code{BRR}, \code{split}, \code{extrapolation}}

\item{k}{number of folds to be defined (only for \code{cv}). Default is \code{k=10}.}

\item{R}{number of times the sample is partioned (needed and used only for \code{cv}, \code{split} or \code{extrapolation} methods). Default R=1.}

\item{B}{number of bootstrap resamples (only for \code{bootstrap} and \code{subbootstrap} methods). Default \code{B=200}.}

\item{cv_error_ind}{method for estimating the error for \code{cv} method. FALSE (default) estimates the error for each test set and defines the cross-validated error as the average of all those errors. Option TRUE estimates the cross-validated error as the weighted average of the loss for each unit}

\item{train_prob}{probability for defining training sets (only for \code{split} and \code{extrapolation} methods)}

\item{method_split}{\code{cv} or \code{bootstrap} (only for \code{split} method)}

\item{seed}{define a seed}
}
\value{
A glmnet object
}
\description{
This function allows as to fit glmnet prediction (linear or 
logistic lasso, ridge, or elastic net) models to complex survey data, 
considering sampling weights in the estimation process and selects the 
lambda that minimizes the error based on different replicating weights 
methods.
}
